Use wrapAndMeasure in BaseRepository
