+++ packages/backend/src/peripherals/database/shared/BaseRepository.ts
@@ -1,4 +1,4 @@
-import { Logger } from '@l2beat/common'
+import { Logger, wrapAndMeasure } from '@l2beat/common'
 import { Knex } from 'knex'
 
 import { Metrics, RepositoryHistogram } from '../../../Metrics'
@@ -201,12 +201,13 @@ export abstract class BaseRepository {
     method: T,
     log: (result: Awaited<ReturnType<T>>) => void,
   ): T {
+    const measured = wrapAndMeasure(method.bind(this), {
+      histogram: this.histogram,
+      labels: { repository: this.constructor.name, method: method.name },
+    })
     /* eslint-disable @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-assignment */
     const fn = async (...args: Parameters<T>) => {
-      const labels = { repository: this.constructor.name, method: method.name }
-      const done = this.histogram.startTimer(labels)
-      const result: Awaited<ReturnType<T>> = await method.call(this, ...args)
-      done(labels)
+      const result: Awaited<ReturnType<T>> = await measured(...args)
       log(result)
       return result
     }
